face anti-spoofing/requirements.txt
albumentations==1.4.10
einops==0.8.0
matplotlib==3.6.3
numpy==1.26.3
Pillow==10.3.0
torch==1.12.0+cu113
torchvision==0.13.0+cu113
tqdm==4.65.2
